clinker-core: extract config leaf-type modules (format, sort, route, utils)
Relocate the self-contained, low-dependency config type clusters out of
the monolithic config/mod.rs into focused submodules, leaving mod.rs a
thinner module-declaration + re-export hub for these clusters:

- format.rs  — InputFormat, OutputFormat, FormatKind, SchemaSource, MergeMode
- sort.rs    — SortField, SortOrder, SortFieldSpec, NullOrder
- route.rs   — RouteMode, RouteBranch, RouteConfig
- utils.rs   — TimeBound, ByteSize, and the shared serde helpers
               (duration / byte-size parsing, default/skip predicates)

Pure relocation: no symbol renames and no public-API change. Every
existing `clinker_core::config::*` path still resolves through the
re-exports. The serde helpers reused by source/output config gain
`pub(crate)` visibility so they remain reachable across the new module
boundary; nothing new becomes `pub`.

Closes #255
