+++ crates/clinker-core/src/config/format.rs
@@ -1,0 +1,137 @@
+//! Input/output format selectors and the schema-source enum.
+
+use super::*;
+use clinker_record::schema_def::SchemaDefinition;
+use serde::de::{self, MapAccess, Visitor};
+use serde::{Deserialize, Deserializer, Serialize};
+
+/// Adjacently tagged format enum for inputs.
+/// `type` selects the format, `options` provides format-specific settings.
+/// `options` is optional — `type: csv` with no `options:` key gives `Csv(None)`.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(tag = "type", content = "options", rename_all = "snake_case")]
+pub enum InputFormat {
+    Csv(Option<CsvInputOptions>),
+    Json(Option<JsonInputOptions>),
+    Xml(Option<XmlInputOptions>),
+    FixedWidth(Option<FixedWidthInputOptions>),
+}
+
+impl InputFormat {
+    /// Short lowercase format name for display.
+    pub fn format_name(&self) -> &'static str {
+        match self {
+            InputFormat::Csv(_) => "csv",
+            InputFormat::Json(_) => "json",
+            InputFormat::Xml(_) => "xml",
+            InputFormat::FixedWidth(_) => "fixed_width",
+        }
+    }
+
+    /// Whether this format has a header row concept (CSV only).
+    pub fn has_header(&self) -> Option<bool> {
+        match self {
+            InputFormat::Csv(Some(opts)) => opts.has_header,
+            _ => None,
+        }
+    }
+}
+
+impl OutputFormat {
+    /// Short lowercase format name for display.
+    pub fn format_name(&self) -> &'static str {
+        match self {
+            OutputFormat::Csv(_) => "csv",
+            OutputFormat::Json(_) => "json",
+            OutputFormat::Xml(_) => "xml",
+            OutputFormat::FixedWidth(_) => "fixed_width",
+        }
+    }
+}
+
+/// Adjacently tagged format enum for outputs.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(tag = "type", content = "options", rename_all = "snake_case")]
+pub enum OutputFormat {
+    Csv(Option<CsvOutputOptions>),
+    Json(Option<JsonOutputOptions>),
+    Xml(Option<XmlOutputOptions>),
+    FixedWidth(Option<FixedWidthOutputOptions>),
+}
+
+/// Supported format types.
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "snake_case")]
+pub enum FormatKind {
+    Csv,
+    Json,
+    Xml,
+    #[serde(rename = "fixed_width")]
+    FixedWidth,
+}
+
+/// Schema source -- file path or inline definition.
+/// Custom Deserialize: YAML string -> FilePath, YAML map -> Inline(SchemaDefinition).
+#[derive(Debug, Clone, Serialize)]
+#[allow(clippy::large_enum_variant)]
+pub enum SchemaSource {
+    FilePath(String),
+    Inline(SchemaDefinition),
+}
+
+impl<'de> Deserialize<'de> for SchemaSource {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        struct SchemaSourceVisitor;
+
+        impl<'de> Visitor<'de> for SchemaSourceVisitor {
+            type Value = SchemaSource;
+
+            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+                formatter
+                    .write_str("a schema file path (string) or an inline schema definition (map)")
+            }
+
+            fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
+            where
+                E: de::Error,
+            {
+                Ok(SchemaSource::FilePath(v.to_owned()))
+            }
+
+            fn visit_map<A>(self, map: A) -> Result<Self::Value, A::Error>
+            where
+                A: MapAccess<'de>,
+            {
+                let def =
+                    SchemaDefinition::deserialize(de::value::MapAccessDeserializer::new(map))?;
+                Ok(SchemaSource::Inline(def))
+            }
+        }
+
+        deserializer.deserialize_any(SchemaSourceVisitor)
+    }
+}
+
+/// Merge ordering discipline across declared inputs.
+///
+/// `Concat` is the deterministic default: each predecessor's records
+/// drain in declaration order, in their per-source FIFO order. Output
+/// is reproducible run-to-run.
+///
+/// `Interleave` reads concurrently from every upstream channel,
+/// emitting records as they arrive. Per-source FIFO is preserved, but
+/// cross-source order follows wall-clock arrival and is therefore
+/// non-deterministic unless `interleave_seed` is set on
+/// [`MergeBody`](crate::config::pipeline_node::MergeBody).
+#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum MergeMode {
+    /// Drain each predecessor sequentially in declaration order.
+    #[default]
+    Concat,
+    /// Drain predecessors concurrently; first-ready-wins.
+    Interleave,
+}
