+++ crates/clinker-core/src/config/route.rs
@@ -1,0 +1,96 @@
+//! Route node fan-out configuration.
+
+use serde::{Deserialize, Serialize};
+
+/// Routing mode for record dispatch.
+#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum RouteMode {
+    /// First-match: evaluate predicates in order, first true wins.
+    #[default]
+    Exclusive,
+    /// All-match: evaluate all predicates, record sent to every matching branch.
+    Inclusive,
+}
+
+/// A named routing branch with a CXL boolean condition.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub struct RouteBranch {
+    pub name: String,
+    pub condition: String,
+}
+
+/// Route configuration for multi-output record dispatch.
+///
+/// Conditions are CXL boolean expressions evaluated per record.
+/// Mandatory `default` prevents silent record drops.
+#[derive(Debug, Clone, Serialize)]
+pub struct RouteConfig {
+    #[serde(default)]
+    pub mode: RouteMode,
+    pub branches: Vec<RouteBranch>,
+    pub default: String,
+}
+
+impl<'de> serde::Deserialize<'de> for RouteConfig {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        #[derive(Deserialize)]
+        struct Raw {
+            #[serde(default)]
+            mode: RouteMode,
+            branches: Vec<RouteBranch>,
+            default: Option<String>,
+        }
+
+        let raw = Raw::deserialize(deserializer)?;
+
+        // Mandatory default
+        let default = raw
+            .default
+            .ok_or_else(|| serde::de::Error::custom("route must have a 'default' output name"))?;
+
+        // Non-empty branches
+        if raw.branches.is_empty() {
+            return Err(serde::de::Error::custom(
+                "route must have at least one branch",
+            ));
+        }
+
+        // Max 256 outputs
+        if raw.branches.len() > 256 {
+            return Err(serde::de::Error::custom(format!(
+                "route has {} branches, maximum is 256",
+                raw.branches.len()
+            )));
+        }
+
+        // Unique branch names
+        let mut seen = std::collections::HashSet::new();
+        for branch in &raw.branches {
+            if !seen.insert(&branch.name) {
+                return Err(serde::de::Error::custom(format!(
+                    "duplicate route branch name '{}'",
+                    branch.name
+                )));
+            }
+        }
+
+        // Default must not collide with a branch name
+        if seen.contains(&default) {
+            return Err(serde::de::Error::custom(format!(
+                "route default '{}' collides with a branch name",
+                default
+            )));
+        }
+
+        Ok(RouteConfig {
+            mode: raw.mode,
+            branches: raw.branches,
+            default,
+        })
+    }
+}
