+++ crates/clinker-core/src/config/sort.rs
@@ -1,0 +1,103 @@
+//! Record-level sort specifications shared by Output and Aggregate.
+
+use serde::de::{self, MapAccess, Visitor};
+use serde::{Deserialize, Deserializer, Serialize};
+
+/// Sort field specification for output and window partition ordering.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub struct SortField {
+    pub field: String,
+    #[serde(default = "default_sort_order")]
+    pub order: SortOrder,
+    /// Null handling during sort. None for output sorting; Some(Last) default for windows.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub null_order: Option<NullOrder>,
+}
+
+/// Sort direction.
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "snake_case")]
+pub enum SortOrder {
+    Asc,
+    Desc,
+}
+
+/// Accepts either a plain string shorthand or a full SortField object in YAML.
+///
+/// Shorthand: `"field_name"` expands to `SortField { field: "field_name", order: Asc, null_order: None }`.
+/// Full: `{ field: "name", order: desc, null_order: first }` deserializes as SortField.
+///
+/// Custom Deserialize: visit_str -> Short, visit_map -> Full.
+/// This gives specific error messages instead of serde(untagged)'s generic "no variant matched".
+#[derive(Debug, Clone, Serialize)]
+pub enum SortFieldSpec {
+    Short(String),
+    Full(SortField),
+}
+
+impl<'de> Deserialize<'de> for SortFieldSpec {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        struct SortFieldSpecVisitor;
+
+        impl<'de> Visitor<'de> for SortFieldSpecVisitor {
+            type Value = SortFieldSpec;
+
+            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+                formatter.write_str("a field name (string) or a sort field object (map with 'field', 'order', 'null_order')")
+            }
+
+            fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
+            where
+                E: de::Error,
+            {
+                Ok(SortFieldSpec::Short(v.to_owned()))
+            }
+
+            fn visit_map<A>(self, map: A) -> Result<Self::Value, A::Error>
+            where
+                A: MapAccess<'de>,
+            {
+                let sf = SortField::deserialize(de::value::MapAccessDeserializer::new(map))?;
+                Ok(SortFieldSpec::Full(sf))
+            }
+        }
+
+        deserializer.deserialize_any(SortFieldSpecVisitor)
+    }
+}
+
+impl SortFieldSpec {
+    /// Resolve to a concrete SortField.
+    pub fn into_sort_field(self) -> SortField {
+        match self {
+            SortFieldSpec::Short(name) => SortField {
+                field: name,
+                order: SortOrder::Asc,
+                null_order: None,
+            },
+            SortFieldSpec::Full(sf) => sf,
+        }
+    }
+}
+
+fn default_sort_order() -> SortOrder {
+    SortOrder::Asc
+}
+
+/// Null handling in sort operations.
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "snake_case")]
+#[derive(Default)]
+pub enum NullOrder {
+    /// Nulls sort before all non-null values.
+    First,
+    /// Nulls sort after all non-null values (SQL convention default).
+    #[default]
+    Last,
+    /// Remove records with null sort keys from the partition.
+    Drop,
+}
